Ajout d'un check de l'existence du fichier PDF

+++ app/controllers/pdf_controller.rb
@@ -7,8 +7,12 @@ class PdfController < ApplicationController
         qcm = get_qcm(params[:id])
 
         filename = "#{Rails.configuration.amc[:pdf_folder]}/#{qcm.id}.pdf"
-        File.open(filename, 'r') do |f|
-          send_data f.read.force_encoding('BINARY'), :filename => qcm.id.to_s, :type => "application/pdf", :disposition => "attachment"
+        if File.exists?(filename)
+          File.open(filename, 'r') do |f|
+            send_data f.read.force_encoding('BINARY'), :filename => qcm.id.to_s, :type => "application/pdf", :disposition => "attachment"
+          end
+        else
+          send_data "", :filename => qcm.id.to_s, :type => "application/pdf", :disposition => "attachment"
         end
     end
 end
